+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/ChinookAppPanel.java
@@ -23,7 +23,6 @@ import is.codion.common.model.UserPreferences;
 import is.codion.common.user.User;
 import is.codion.framework.demos.chinook.domain.Chinook;
 import is.codion.framework.demos.chinook.model.ChinookAppModel;
-import is.codion.framework.demos.chinook.model.EmployeeEditModel;
 import is.codion.swing.common.ui.component.combobox.Completion;
 import is.codion.swing.common.ui.component.table.FilteredTable;
 import is.codion.swing.common.ui.component.table.FilteredTableCellRenderer;
@@ -126,13 +125,11 @@ public final class ChinookAppPanel extends EntityApplicationPanel<ChinookAppMode
 
     SwingEntityModel.Builder employeeModelBuilder =
             SwingEntityModel.builder(Employee.TYPE)
-                    .detailModel(SwingEntityModel.builder(Customer.TYPE))
-                    .editModel(EmployeeEditModel.class);
+                    .detailModel(SwingEntityModel.builder(Customer.TYPE));
 
     EntityPanel.Builder employeePanelBuilder =
             EntityPanel.builder(employeeModelBuilder)
                     .editPanel(EmployeeEditPanel.class)
-                    .tablePanel(EmployeeTablePanel.class)
                     .detailPanel(customerPanelBuilder)
                     .layout(detailPanelState(PanelState.HIDDEN))
                     .preferredSize(new Dimension(1000, 500));